Added support for csv-files. Due do new software on the Aquadas measurement instrument on some ships, the output file format has changed. The naming of the files is different from the dat-files, while also the content if the file is different: 	- No description in the header of the file with information about the measurement 	- Different coördinate system, WGS84 instead of RD New 	- Sensor height sometimes in [cm], sometimes in [m], 	  sometimes all values positive, sometimes negative
This lead to all kinds of changes, because most of the logic was built around the format of the dat-files.
The dat-files will still be in use, because not all ships / measurement instruments are running the new software.
It is also expected that the file format will change again, because this is not the final version, but this could
take a while, and the software project to do this is on hold at the moment. So it had to work with this kind of
intermediate version for now.

Also included in this commit is the processing of a number of new measurements.